requirements.txt
Django==2.0
django-cors-headers==3.2.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-jet==1.0.8
django-pdb==0.6.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Markdown==3.1
npm==0.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
xlwt==1.3.0
